+++ NPCMapLocations/docs/release-notes.md
@@ -3,8 +3,8 @@
 # Release notes
 ## Upcoming version
 * Improved the new `npc_map_locations_summary` command:
-  * Added a list of unknown locations.
-  * It now shows built-in map vectors too.
+  * Added a table of location data.
+  * The map vectors list now includes those bundled with NPC Map Locations.
 * Fixed farm buildings not shown on load if any location without map vectors is also excluded.
 
 ## 2.7.0